requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
gradio>=3.0.0
google-generativeai>=0.3.0
torch>=1.9.0
torchvision>=0.10.0
python-dotenv>=1.0.0
scikit-image>=0.19.0 